check for old output and delete

+++ L1_task-trust_model-01.sh
@@ -10,7 +10,9 @@ DATA=${maindir}/fmriprep/fmriprep/sub-${sub}/func/sub-${sub}_task-trust_run-0${r
 MAINOUTPUT=${maindir}/fsl/sub-${sub}/
 mkdir -p $MAINOUTPUT
 OUTPUT=${MAINOUTPUT}/L1_task-trust_model-01_type-act_run-0${run}
-rm -rf ${OUTPUT}.feat
+if [ ! -e ${OUTPUT}.feat/cluster_mask_zstat1.nii.gz ]; then
+	rm -rf ${OUTPUT}.feat
+fi
 
 MISSED_TRIAL=${EVDIR}_missed_trial.txt
 if [ -e $MISSED_TRIAL ]; then
